requirements.txt
streamlit
sentence-transformers
transformers
spacy
pypdf2
speechrecognition
pyttsx3
google-generativeai
google-auth

streamlit-webrtc
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
